+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes in keycloak-group-management will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.0] - 2024-05-23
+
+### Added
+- SSH Public Keys Page (UI)
+- Error Handling Group Admin Actions
+
+### Fixed
+- Footer CSS
+- Membership Expiration Input Bug
+
 ## [0.17.0] - 2024-05-16
 
 ### Fixed